Same late clock in and late clock out bug fix
There was a problem when someone clocked in late the same amount of time that the clocked out late. It was considering them a no call no show. Added a check to calculate the actual number of hours they were scheduled that then compares itself to the potential no call no show hours.