Spanish Option Added
Added a Spanish option as well as removed bug introduced by last patch that caused no missed shift text to be generated. Had to add a separate method for converting to military time so that we can accurately compare the late shifts to the actual length of the shift divided by two. If they are equal, then we know it is a missed shift.