Multiple Lates Bug Fix
Fixed a bug introduced when I made the "convert_first_name_last_name()" function. I forgot to add it to an if statement in the function that creates the dictionary of lates. This meant each person could only have one recorded and the last one would always overwrite the others. Now each individual can have any number of documents.